+++ src/joedb/journal/iostream.h
@@ -1,18 +1,18 @@
 #ifndef joedb_iostream_declared
 #define joedb_iostream_declared
 
-#include "joedb/journal/streambuf.h"
+#include "joedb/journal/filebuf.h"
 
 #include <iostream>
 
 namespace joedb
 {
  /// @ingroup journal
- class iostream: private joedb::streambuf, public std::iostream
+ class iostream: private joedb::filebuf, public std::iostream
  {
   public:
    iostream(joedb::Abstract_File &file):
-    streambuf(file),
+    filebuf(file),
     std::iostream(static_cast<std::streambuf *>(this))
    {
    }