+++ test/compiler/Interpreted_Database.cpp
@@ -47,7 +47,7 @@ TEST(Compiler, Interpreted_Stream_Error_Alternative)
    "this_file_does_not_exist.joedbi",
    joedb::Open_Mode::read_existing
   );
-  joedb::streambuf buf(file);
+  joedb::filebuf buf(file);
   std::istream is(&buf);
   FAIL() << "Expected exception";
  }